+++ NEWS.md
@@ -5,6 +5,7 @@
 * Improve message feedback for downloading objects
 * `gcs_upload()` will now retry with a warning if using legacy ACL settings for a bucket with uniform bucket-level access enabled - will retry with `predefinedAcl='bucketLevel'`
 * Add `gcs_parse_rds()` to help parsing RDS objects for `gcs_get_object()` #146
+* Add ability to request noncurrent versions of objects in `gcs_get_object()` via `generation` parameter.
 
 # googleCloudStorageR 0.6.0